+++ src/main/java/com/example/home_work_49/models/WorkExperienceInfo.java
@@ -1,9 +1,9 @@
 package com.example.home_work_49.models;
 
 public class WorkExperienceInfo {
-    private int id;
-    private int resumeId;
-    private int years;
+    private Long id;
+    private Long resumeId;
+    private Integer years;
     private String companyName;
     private String position;
     private String responsibilities;
